Optimize `Arena::reserve`: Replace one rel & acq with downgrade
This would avoid the release of the write guard and reacquire of the
upgradable read guard, which might involve one syscall for waking up all
readers or one writer and another syscall to wait for
upgradable reader/writer to finish to acquire the
upgradable read guard, along with several atomic operations.

There is also possiblity for the other thread to acquire the upgradable
read guard, causing them to be wake up.

This is less efficient than executing the pushing on the same thread,
since it might involve cache misses and context switching.

Using the downgrade, there can be at most one syscall issues to wake up
all the readers and the thread will continue to add more buckets,
improving the cache hits and avoid context switching.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>