Bugfix: flag code that only works in debug mode
This fixes a bug that breaks release builds.

There a `slice.is_same_arc` call in a debug assert. The `is_same_arc`
method is behind the `debug_assertions` flag, which compiles in debug
mode, but not in release mode. The `debug_assert!` in `Arcs::do_grow`
contained a `is_same_arc` call, but `debug_assert` is was still being
compiled in release, so it failed.